refactor: fixed issues in joining processors

+++ src/modalities/dataloader/preprocessing/queued_processing/process_controller.py
@@ -11,6 +11,7 @@ from modalities.utils.logging import get_logger
 @dataclass
 class PipelineStep:
     name: str
+    poisonable: bool
     input_queue: mp.Queue
     processors: list[Processor]
 
@@ -30,14 +31,22 @@ class ProcessController:
         If the stop_event is set, the processors are terminated
         """
         # poison the input queues of the processors
-        for _ in tqdm.tqdm(step.processors, desc=f"Poisoning {step.name} processes"):
-            if step.input_queue is not None:
-                step.input_queue.put(None)
+        if step.poisonable:
+            for _ in tqdm.tqdm(step.processors, desc=f"Poisoning {step.name} processes"):
+                if step.input_queue is not None:
+                    step.input_queue.put(None)
 
         # join the processors
         num_exits = 0
         while num_exits < len(step.processors):
             processor = step.processors[num_exits]
+
+            # if the processor is not alive, we continue with the next one
+            if not processor.is_alive():
+                get_logger().info(f"Processor {processor.full_name} is not alive. Continuing with the next processor.")
+                num_exits += 1
+                continue
+            # if the stop event is set, we terminate the processor
             if self._stop_event.is_set():
                 try:
                     processor.terminate()
@@ -51,12 +60,14 @@ class ProcessController:
                     continue
                 get_logger().info(f"Terminated processor {processor.full_name}")
                 num_exits += 1
+            # if the stop event is not set, we join the processor
             else:
-                try:
-                    processor.join(timeout=self.join_timeout)
-                except TimeoutError:
+                get_logger().info(f"Joining {processor.full_name} ...")
+                processor.join(timeout=self._join_timeout)
+                if processor.exitcode is None:
+                    get_logger().info(f"Joining {processor.full_name} timed out. Exit code: {processor.exitcode} ...")
                     continue
-                get_logger().info(f"Joined processor {processor.full_name}")
+                get_logger().info(f"Joined processor {processor.full_name}. Exit code: {processor.exitcode}")
                 num_exits += 1
 
     def run(self):
